refactor: extract account actions from App
